refactor(freepsflux): flux to new operator interface and optional compilation
