+++ Makefile
@@ -19,7 +19,7 @@ build/freepsd: build connectors/http/static_server_content/chota.min.css
 	go build -ldflags="-X ${PACKAGE}/utils.Version=${VERSION} -X ${PACKAGE}/utils.CommitHash=${COMMIT_HASH} -X ${PACKAGE}/utils.BuildTime=${BUILD_TIMESTAMP} -X ${PACKAGE}/utils.Branch=${BRANCH}" -o build/freepsd freepsd/freepsd.go
 
 build/freepsd-light: build connectors/http/static_server_content/chota.min.css
-	go build -tags nopostgress,nomuteme,nobluetooth -ldflags="-X ${PACKAGE}/utils.Version=${VERSION} -X ${PACKAGE}/utils.CommitHash=${COMMIT_HASH} -X ${PACKAGE}/utils.BuildTime=${BUILD_TIMESTAMP} -X ${PACKAGE}/utils.Branch=${BRANCH}" -o build/freepsd-light freepsd/freepsd.go
+	go build -tags nopostgress,nomuteme,nobluetooth,noinflux -ldflags="-X ${PACKAGE}/utils.Version=${VERSION} -X ${PACKAGE}/utils.CommitHash=${COMMIT_HASH} -X ${PACKAGE}/utils.BuildTime=${BUILD_TIMESTAMP} -X ${PACKAGE}/utils.Branch=${BRANCH}" -o build/freepsd-light freepsd/freepsd.go
 
 # if you are reading this to learn how freepsd is deployed: freepsd runs without any additional libraries or setup. Just run it.
 # this just creates a user and a service and an optional update-script (that should only be used if you want to automatically build from source)
