+++ connectors/freepsflux/freepsflux_test.go
@@ -1,3 +1,5 @@
+//go:build !noinflux
+
 package freepsflux
 
 import (
