+++ freepsd/freepsd.go
@@ -58,6 +58,7 @@ func mainLoop() bool {
 	availableOperators := []base.FreepsOperator{
 		&freepsbluetooth.Bluetooth{},
 		&muteme.MuteMe{},
+		&freepsflux.OperatorFlux{},
 	}
 
 	logger := logrus.StandardLogger()
@@ -92,7 +93,6 @@ func mainLoop() bool {
 	}
 	ge.AddOperator(mqtt.NewMQTTOp(cr))
 	ge.AddOperator(telegram.NewTelegramOp(cr))
-	ge.AddOperator(freepsflux.NewFluxMod(cr))
 	ge.AddOperator(wled.NewWLEDOp(cr))
 	ge.AddOperator(ui.NewHTMLUI(cr, ge))
 	ge.AddOperator(fritz.NewOpFritz(cr))
